[#0] TF : Adding basic file move feature of directorytree, and implementing basic text edit tool

+++ src/experiments/directorytree/factories/FileSystemTreeViewFactory.java
@@ -1,18 +1,21 @@
 package experiments.directorytree.factories;
 
-import experiments.directorytree.Main;
-import experiments.directorytree.Util;
+import experiments.directorytree.prompt.GeneralPrompt;
+import experiments.directorytree.utils.FileUtils;
 import experiments.directorytree.prompt.FilePrompt;
 import experiments.directorytree.singletons.SImages;
 import experiments.directorytree.tree_views.FileSystemTreeView;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.scene.control.*;
-import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.scene.input.*;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.util.Objects;
 
 //http://www.drdobbs.com/jvm/a-javafx-text-editor-part-1/240142297
@@ -37,11 +40,14 @@ public class FileSystemTreeViewFactory {
                         setTooltip(null);
                         setContextMenu(null);
                         setGraphic(null);
+                        clearDragEvents(this);
+
                     } else {
                         setText(createText(this));
                         setTooltip(createTooltip(this));
                         setContextMenu(createContextMenu(this));
                         setGraphic(createGraphic(this));
+                        setOnDragEvents(this);
                     }
                 }
             };
@@ -119,13 +125,17 @@ public class FileSystemTreeViewFactory {
             @Override
             public void handle(ActionEvent event) {
                 if (treeItemFile.exists()) {
-                    boolean deleted = treeItemFile.delete();
+                    boolean confirmation = GeneralPrompt.showAndGetConfirmation(String.format("Are you sure you want to delete %s?", fileName));
 
-                    if (deleted == true) {
-                        FileSystemTreeView fileSystemTreeView = (FileSystemTreeView) cell.getTreeView();
-                        refresh(fileSystemTreeView);
-                    } else {
-                        throw new RuntimeException("Could not delete file");
+                    if (confirmation) {
+                        boolean deleted = treeItemFile.delete();
+
+                        if (deleted) {
+                            FileSystemTreeView fileSystemTreeView = (FileSystemTreeView) cell.getTreeView();
+                            refresh(fileSystemTreeView);
+                        } else {
+                            throw new RuntimeException("Could not delete file");
+                        }
                     }
                 }
             }
@@ -170,7 +180,7 @@ public class FileSystemTreeViewFactory {
             if (file.isDirectory()) {
                 imageView.setImage(SImages.getInstance().getImage("blue-folder.png"));
             } else {
-                String extension = Util.getFileExtension(file);
+                String extension = FileUtils.getFileExtension(file);
 
                 if (Objects.equals(extension, ".txt")) {
                     imageView.setImage(SImages.getInstance().getImage("file.png"));
@@ -183,6 +193,100 @@ public class FileSystemTreeViewFactory {
         }
 
         return imageView;
+    }
+
+    protected static void setOnDragEvents(TreeCell<File> cell) {
+        TreeItem<File> treeItem = cell.getTreeItem();
+        File file = treeItem.getValue();
+
+        cell.setOnDragDetected(new EventHandler<MouseEvent>() {
+            public void handle(MouseEvent event) {
+                /* drag was detected, start drag-and-drop gesture*/
+                System.out.println("onDragDetected");
+
+                /* allow any transfer mode */
+                Dragboard dragboard = cell.startDragAndDrop(TransferMode.COPY_OR_MOVE);
+
+                /* put a string on dragboard */
+                ClipboardContent content = new ClipboardContent();
+                content.putString(file.getAbsolutePath());
+                dragboard.setContent(content);
+
+                event.consume();
+            }
+        });
+
+        cell.setOnDragOver(new EventHandler<DragEvent>() {
+            public void handle(DragEvent event) {
+                try {
+                    if (validDropFromDrag(cell, event)) {
+                        event.acceptTransferModes(TransferMode.COPY_OR_MOVE);
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+
+                event.consume();
+            }
+        });
+
+        cell.setOnDragDropped(new EventHandler<DragEvent>() {
+            public void handle(DragEvent event) {
+                /* data dropped */
+                System.out.println("onDragDropped");
+                /* if there is a string data on dragboard, read it and use it */
+                try {
+                    if (validDropFromDrag(cell, event)) {
+                        String sourceAbsolutePath = event.getDragboard().getString();
+                        File sourceFile = Paths.get(sourceAbsolutePath).toFile();
+                        try {
+                            Files.move(sourceFile.toPath(), file.toPath().resolve(sourceFile.getName()), StandardCopyOption.REPLACE_EXISTING);
+                            FileSystemTreeView fileSystemTreeView = (FileSystemTreeView) cell.getTreeView();
+                            refresh(fileSystemTreeView);
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+                /* let the source know whether the string was successfully
+                 * transferred and used */
+                event.setDropCompleted(true);
+
+                event.consume();
+            }
+        });
+
+        System.out.println("Added mouse event");
+    }
+
+    protected static void clearDragEvents(TreeCell<File> cell) {
+        cell.setOnDragDetected(null);
+        cell.setOnDragOver(null);
+        cell.setOnDragEntered(null);
+        cell.setOnDragExited(null);
+        cell.setOnDragDropped(null);
+        cell.setOnDragDone(null);
+    }
+
+    protected static boolean validDropFromDrag(TreeCell<File> cell, DragEvent event) throws IOException {
+        TreeItem<File> treeItem = cell.getTreeItem();
+        File destFile = treeItem.getValue();
+        boolean validDrop = false;
+                /* data is dragged over the target */
+
+                /* accept it only if it is  not dragged from the same node
+                 * and if it has a string data */
+        if (event.getGestureSource() != cell &&
+                event.getDragboard().hasString()) {
+            String sourceAbsolutePath = event.getDragboard().getString();
+            File sourceFile = Paths.get(sourceAbsolutePath).toFile();
+
+            validDrop = FileUtils.isValidFileMove(sourceFile, destFile);
+        }
+
+        return validDrop;
     }
 }
